# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/tenants": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Lists tenants",
                "operationId": "listTenants",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantPagingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Adds a new tenant",
                "operationId": "storeTenant",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/tenants/{id}": {
            "get": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Get Tenant information",
                "description": "Returns Tenant data",
                "operationId": "showTenant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Update Tenant information",
                "description": "Returns Tenant data",
                "operationId": "updateTenant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tenants"
                ],
                "summary": "Delete Tenant",
                "description": "Returns deleted Tenant data",
                "operationId": "destroyTenant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tenant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TenantResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BaseModel": {
                "title": "BaseModel",
                "description": "BaseModel",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BaseModel"
                }
            },
...
            "NamedBaseModel": {
                "title": "NamedBaseModel",
                "description": "NamedBaseModel",
                "type": "object",
                "xml": {
                    "name": "NamedBaseModel"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseModel"
                    },
                    {
                        "properties": {
                            "name": {
                                "title": "Name",
                                "description": "Name of the entity",
                                "type": "string",
                                "example": "Elmofuchto"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "Tenant": {
                "title": "Tenant",
                "description": "Tenant model",
                "xml": {
                    "name": "Tenant"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/NamedBaseModel"
                    }
                ]
            },
...
            "TenantStoreRequest": {
                "title": "TenantStoreRequest",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new Tenant",
                        "type": "string",
                        "example": "A nice Tenant"
                    }
                },
                "type": "object"
            },
            "TenantUpdateRequest": {
                "title": "TenantUpdateRequest",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Update name of the new Tenant",
                        "type": "string",
                        "example": "A nice Tenant"
                    }
                },
                "type": "object"
            },
...
            "Meta": {
                "title": "Meta",
                "properties": {
                    "current_page": {
                        "title": "Current Page",
                        "description": "Current Page of the data",
                        "type": "integer"
                    },
                    "from": {
                        "title": "From",
                        "description": "From",
                        "type": "integer"
                    },
                    "last_page": {
                        "title": "Last Page",
                        "description": "Last Page of the data",
                        "type": "integer"
                    },
                    "per_page": {
                        "title": "Per Page",
                        "description": "Per Page of the data",
                        "type": "integer"
                    },
                    "to": {
                        "title": "To",
                        "description": "To",
                        "type": "integer"
                    },
                    "total": {
                        "title": "Total",
                        "description": "Total",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "TenantPagingResource": {
                "title": "TenantPagingResource",
                "description": "TenantPagingResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tenant"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Meta"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TenantPagingResource"
                }
            },
            "TenantResource": {
                "title": "TenantResource",
                "description": "TenantResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tenant"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TenantResource"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format Bearer 4|oeXad4kChJT43wli90LOd1VbFhtuGuEdvxvEHMAtcb025185",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}